adk_mcp/requirements.txt
fastapi
uvicorn
python-multipart
beautifulsoup4
requests
google-cloud-bigquery[bqstorage,pandas]
google-cloud-storage
google-cloud-aiplatform
google-adk
langchain
langchain-google-genai
python-dotenv
pandas
openpyxl
aiohttp
asyncio
httpx
lxml
html5lib
yfinance
OpenDartReader
google-cloud-secret-manager
